Analysis

BialkaliSpectrum.dressed_dipole_moment

BialkaliSpectrum.dressed_dipole_moment_vector

BialkaliSpectrum.with_ground_state

Types

Methods

BialkaliSpectrum.dressed_dipole_moment — Method

dressed_dipole_moment(
    hamiltonian_parts::HamiltonianParts,
    g,
    e,
    polarization::Int,
)::ComplexF64

BialkaliSpectrum.dressed_dipole_moment_vector — Method

dressed_dipole_moment_vector(
    hamiltonian_parts::HamiltonianParts,
    g,
    e,
    polarization::Int,
)::Vector{ComplexF64}

BialkaliSpectrum.with_ground_state — Method

with_ground_state(
    spectra,
    ground_basis_state::State;
    groupby=:fields
)

Returns a new DataFrame with the contents of spectra, plus an additional column :ground_state containing the :index of the state with the highest overlap with ground_basis_state, in the same group determined by groupby.

Used internally but not necessary probably??