Fields

BialkaliSpectrum.SphericalUnitVector

BialkaliSpectrum.T⁽²⁾

BialkaliSpectrum.T⁽¹⁾

BialkaliSpectrum.UnitVectorX

BialkaliSpectrum.UnitVectorY

BialkaliSpectrum.UnitVectorZ

BialkaliSpectrum.tensor_dot

Types

BialkaliSpectrum.SphericalUnitVector — Type

SphericalUnitVector(magnitude, θ, φ)
SphericalUnitVector(v::SphericalVector)

Construct a unit vector with polar angle θ, and azimuthal angle φ.

Represents the direction of an external field vector in spherical coordinates. Currently, SphericalUnitVectors may be negated but no other mathematical operations are implemented.

Vectors along $x$, $y$, or $z$ can be quickly constructed using UnitVectorX, UnitVectorY, and UnitVectorZ, respectively.

See also SphericalVector, T⁽¹⁾, T⁽²⁾.

Methods

BialkaliSpectrum.T⁽²⁾ — Method

T⁽²⁾(v)

Construct the components of the rank 2 spherical tensor $T⁽²⁾(v, v)$.

BialkaliSpectrum.T⁽¹⁾ — Method

T⁽¹⁾(v)

Construct the components of the rank 1 spherical tensor $T⁽¹⁾(v)$.

BialkaliSpectrum.UnitVectorX — Method

UnitVectorX()

Construct a SphericalUnitVector along x.

BialkaliSpectrum.UnitVectorY — Method

UnitVectorY()

Construct a SphericalUnitVector along y.

BialkaliSpectrum.UnitVectorZ — Method

UnitVectorZ()

Construct a SphericalUnitVector along z.

BialkaliSpectrum.tensor_dot — Method

tensor_dot(a, b)

Contract two spherical tensors a and b.