Hamiltonian

BialkaliSpectrum.h_diagonal

BialkaliSpectrum.h_rank_0

BialkaliSpectrum.h_rank_1

BialkaliSpectrum.h_rank_2

BialkaliSpectrum.h_tensor_component

Types

Methods

BialkaliSpectrum.h_diagonal — Method

h_diagonal(basis, matrix_element)

Create a diagonal matrix where the (i, i)th element is matrix_element(basis[i], basis[i]).

BialkaliSpectrum.h_rank_0 — Method

h_rank_0(basis, matrix_element)

Create a matrix where the (i, j)th element is matrix_element(basis[j], basis[i]).

Note that matrix_element must have the signature (State, State) -> ComplexF64.

BialkaliSpectrum.h_rank_1 — Method

h_rank_1(basis, matrix_element)

Create a vector of 3 matrices to contract against a rank 1 spherical tensor.

Note that matrix_element must have the signature (Int, State, State) -> ComplexF64. The first argument is the tensor component p in -1:1.

BialkaliSpectrum.h_rank_2 — Method

h_rank_2(basis, matrix_element)

Create a vector of 5 matrices to contract against a rank 2 spherical tensor.

Note that matrix_element must have the signature (Int, State, State) -> ComplexF64. The first argument is the tensor component p in -2:2.

BialkaliSpectrum.h_tensor_component — Method

h_tensor_component(basis, matrix_element, p)

Create a matrix where the (i, j)th element is matrix_element(p, basis[j], basis[i]).

Note that matrix_element must have the signature (Int, State, State) -> ComplexF64. The first argument is the tensor component p.